refactor: move dayoffset methods
