+++ AnkiDroid/src/test/java/com/ichi2/anki/preferences/PreferencesTest.kt
@@ -22,8 +22,6 @@ import androidx.test.core.app.ActivityScenario
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.ichi2.anki.R
 import com.ichi2.anki.RobolectricTest
-import com.ichi2.anki.preferences.Preferences.Companion.getDayOffset
-import com.ichi2.anki.preferences.Preferences.Companion.setDayOffset
 import com.ichi2.libanki.exception.ConfirmModSchemaException
 import com.ichi2.preferences.HeaderPreference
 import com.ichi2.testutils.getJavaMethodAsAccessible
